# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/cities/": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a City",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Create a City",
                "parameters": [
                    {
                        "description": "Create a City",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_dto.CreateCityRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "City response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_dto.CityResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/cities/get-by-filter": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get Cities",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Get Cities",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_dto.PaginationInputWithFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_dto.PagedList-github_com_paghapour_golang-clean-web-api_api_dto_CityResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/cities/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get a City",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Get a City",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_dto.CityResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update a City",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Update a City",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update a City",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_dto.UpdateCityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_dto.CityResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Delete a City",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Delete a City",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_paghapour_golang-clean-web-api_api_dto.CityResponse": {
            "type": "object",
            "properties": {
                "country": {
                    "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_dto.CountryResponse"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_paghapour_golang-clean-web-api_api_dto.CountryResponse": {
            "type": "object",
            "properties": {
                "cities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_dto.CityResponse"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_paghapour_golang-clean-web-api_api_dto.CreateCityRequest": {
            "type": "object",
            "required": [
                "countryId",
                "name"
            ],
            "properties": {
                "countryId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                }
            }
        },
...
        "github_com_paghapour_golang-clean-web-api_api_dto.Filter": {
            "type": "object",
            "properties": {
                "filterType": {
                    "description": "text number",
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "type": {
                    "description": "contains notContains equals notEqual startsWith lessThan lessThanOrEqual greaterThan greaterThanOrEqual inRange endsWith",
                    "type": "string"
                }
            }
        },
...
        "github_com_paghapour_golang-clean-web-api_api_dto.PagedList-github_com_paghapour_golang-clean-web-api_api_dto_CityResponse": {
            "type": "object",
            "properties": {
                "hasNextPage": {
                    "type": "boolean"
                },
                "hasPreviousPage": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_dto.CityResponse"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalRows": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_paghapour_golang-clean-web-api_api_dto.PaginationInputWithFilter": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_dto.Filter"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "sort": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_dto.Sort"
                    }
                }
            }
        },
...
        "github_com_paghapour_golang-clean-web-api_api_dto.Sort": {
            "type": "object",
            "properties": {
                "colId": {
                    "type": "string"
                },
                "sort": {
                    "type": "string"
                }
            }
        },
        "github_com_paghapour_golang-clean-web-api_api_dto.UpdateCityRequest": {
            "type": "object",
            "properties": {
                "countryId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                }
            }
        },
...
        "github_com_paghapour_golang-clean-web-api_api_helper.BaseHttpResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "result": {},
                "resultCode": {
                    "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_helper.ResultCode"
                },
                "success": {
                    "type": "boolean"
                },
                "validationErrors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_paghapour_golang-clean-web-api_api_validations.ValidationError"
                    }
                }
            }
        },
        "github_com_paghapour_golang-clean-web-api_api_helper.ResultCode": {
            "type": "integer",
            "enum": [
                0,
                40001,
                40101,
                40301,
                40401,
                42901,
                42902,
                50001,
                50002
            ],
            "x-enum-varnames": [
                "Success",
                "ValidationError",
                "AuthError",
                "ForbiddenError",
                "NotFoundError",
                "LimiterError",
                "OtpLimiterError",
                "CustomRecovery",
                "InternalError"
            ]
        },
        "github_com_paghapour_golang-clean-web-api_api_validations.ValidationError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "property": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AuthBearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}